refactor: 기존 컴포넌트를 ResponseActionList로 변경, chats widget 폴더 생성
