+++ lib/widgets/chats/response_action_list.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class ResponseActionList extends StatelessWidget {
+  final List<Widget> children;
+  final double spacing;
+  final CrossAxisAlignment crossAxisAlignment;
+
+  const ResponseActionList({
+    Key? key,
+    required this.children,
+    this.spacing = 8.0,
+    this.crossAxisAlignment = CrossAxisAlignment.start,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: crossAxisAlignment,
+      children: [
+        for (int i = 0; i < children.length; i++) ...[
+          if (i > 0) SizedBox(height: spacing),
+          children[i],
+        ],
+      ],
+    );
+  }
+}
